+++ js/app/stores/UserStore.js
@@ -1,0 +1,154 @@
+var btoa = require('btoa');
+var reqwest = require('reqwest');
+var assign = require('object-assign');
+var EventEmitter = require('events').EventEmitter;
+
+var KanbanerDispatcher = require('../dispatcher/KanbanerDispatcher');
+var KanbanerConstants = require('../constants/KanbanerConstants');
+
+var ActionTypes = KanbanerConstants.ActionTypes;
+
+
+var USER_LOGGED_IN = 'user-logged-in';
+var USER_LOGGED_OUT = 'user-logged-out';
+var USER_AUTHENTICATION_FAILED = 'user-authentication-failed';
+
+var user = null;
+var token = null;
+
+/**
+ * Tries to authenticate user with login/password
+ * @param {string} username
+ * @param {password} password
+ */
+var authenticateUser = function(username, password) {
+  var authToken = btoa(username + ":" + password)
+  reqwest({
+    url: 'https://api.github.com/user',
+    method: 'get',
+    headers: {
+      "Authorization": "Basic " + authToken
+    }
+  })
+    .then(function(response) {
+      user = response;
+      token = authToken;
+      UserStore.emitUserLoggedIn();
+    })
+    .fail(function(error, msg) {
+      UserStore.emitAuthenticationFailed();
+    });
+};
+
+/**
+ * Clear user, token variables emit USER_LOGGED_OUT event
+ */
+var logoutUser = function() {
+  user = null;
+  token = null;
+  UserStore.emitUserLoggedOut();
+};
+
+var UserStore = assign({}, EventEmitter.prototype, {
+  /**
+   * Returns is user logged in or not
+   * @returns {boolean}
+   */
+  isUserLoggedIn: function() {
+    return (user)?true:false;
+  },
+
+  /**
+   * Returns user
+   */
+  getUser: function() {
+    return user;
+  },
+
+  /**
+   * Add listener for user logged in event
+   * @param {function} callback
+   */
+  addUserLoggedInListener: function(callback) {
+    this.on(USER_LOGGED_IN, callback);
+  },
+
+  /**
+   * Remove listener for user logged out event
+   * @param {function} callback
+   */
+  removeUserLoggedInListener: function(callback) {
+    this.removeListener(USER_LOGGED_IN, callback);
+  },
+
+  /**
+   * User logged in event;
+   */
+  emitUserLoggedIn: function() {
+    this.emit(USER_LOGGED_IN);
+  },
+
+  /**
+   * Add callback for user logged out event
+   * @param {function} callback
+   */
+  addUserLoggedOutListener: function(callback) {
+    this.on(USER_LOGGED_OUT, callback)
+  },
+
+  /**
+   * Remove callback for user logged out event
+   * @param {function} callback
+   */
+  removeUserLoggedOutListener: function(callback) {
+    this.removeListener(USER_LOGGED_OUT, callback)
+  },
+
+  /**
+   * User logged out event
+   */
+  emitUserLoggedOut: function() {
+    this.emit(USER_LOGGED_OUT);
+  },
+
+  /**
+   * Add authentication failed event listener
+   * @param {function} callback
+   */
+  addAuthenticationFailedListener: function(callback) {
+    this.on(USER_AUTHENTICATION_FAILED, callback);
+  },
+
+  /**
+   * Remove authentaction failed event listener
+   * @param {function} callback
+   */
+  removeAuthenticationFailedListener: function(callback) {
+    this.removeListener(USER_AUTHENTICATION_FAILED, callback);
+  },
+
+  /**
+   * Send authentication failed event
+   */
+  emitAuthenticationFailed: function() {
+    this.emit(USER_AUTHENTICATION_FAILED);
+  }
+});
+
+UserStore.dispatcherToken = KanbanerDispatcher.register(function(payload) {
+  var action = payload.action;
+
+  console.log(action.type);
+  switch(action.type) {
+    case ActionTypes.SEND_AUTHENTICATION_CREDENTIALS:
+      console.log("Send auth credentials");
+      authenticateUser(action.username, action.password);
+      break;
+    case ActionTypes.USER_LOGOUT:
+      console.log("Logout user");
+      logoutUser();
+      break;
+  }
+});
+
+module.exports = UserStore;